modular the settings manager
