+++ Code/renderer/settingsManager/themes.js
@@ -1,0 +1,231 @@
+const FULL_THEMES = {
+  'navy-dark': {
+    label: 'Navy Dark', emoji: '\u{1F30A}', dark: true,
+    bg: { base:'#070d1a', surface:'#0c1427', elevated:'#111d34', overlay:'#162038',
+          hover:'rgba(255,255,255,0.04)', active:'rgba(255,255,255,0.07)',
+          raised:'#1a2540', statusbar:'#0e1830', tree:'#0a1220' },
+    border: { subtle:'rgba(255,255,255,0.06)', default:'rgba(255,255,255,0.10)',
+              strong:'rgba(255,255,255,0.18)', mid:'rgba(255,255,255,0.12)' },
+    text: { primary:'#eef2ff', secondary:'#94a3c4', muted:'#556080', faint:'#364060' },
+    accent: '#f0b429',
+    green:'#34d399', red:'#f87171', blue:'#60a5fa', purple:'#a78bfa', yellow:'#fbbf24',
+    depths: ['#f0b429','#38bdf8','#a78bfa','#2dd4bf','#fb7185'],
+  },
+  'cream-light': {
+    label: 'Cream Light', emoji: '\u2601\uFE0F', dark: false,
+    bg: { base:'#f0f4f8', surface:'#ffffff', elevated:'#f8fafc', overlay:'#eef2f7',
+          hover:'rgba(0,0,0,0.04)', active:'rgba(0,0,0,0.07)',
+          raised:'#e8eef6', statusbar:'#f1f5fb', tree:'#f4f7fb' },
+    border: { subtle:'rgba(0,0,0,0.06)', default:'rgba(0,0,0,0.10)',
+              strong:'rgba(0,0,0,0.18)', mid:'rgba(0,0,0,0.12)' },
+    text: { primary:'#0f172a', secondary:'#334155', muted:'#64748b', faint:'#94a3b8' },
+    accent: '#0d9488',
+    green:'#059669', red:'#dc2626', blue:'#2563eb', purple:'#7c3aed', yellow:'#d97706',
+    depths: ['#0d9488','#0284c7','#7c3aed','#059669','#be123c'],
+  },
+  'catppuccin-mocha': {
+    label: 'Catppuccin Mocha', emoji: '\u{1F431}', dark: true,
+    bg: { base:'#1e1e2e', surface:'#181825', elevated:'#1e1e2e', overlay:'#24273a',
+          hover:'rgba(205,214,244,0.05)', active:'rgba(205,214,244,0.08)',
+          raised:'#313244', statusbar:'#181825', tree:'#1e1e2e' },
+    border: { subtle:'rgba(205,214,244,0.08)', default:'rgba(205,214,244,0.13)',
+              strong:'rgba(205,214,244,0.22)', mid:'rgba(205,214,244,0.10)' },
+    text: { primary:'#cdd6f4', secondary:'#bac2de', muted:'#6c7086', faint:'#45475a' },
+    accent: '#cba6f7',
+    green:'#a6e3a1', red:'#f38ba8', blue:'#89b4fa', purple:'#cba6f7', yellow:'#f9e2af',
+    depths: ['#cba6f7','#89b4fa','#a6e3a1','#f9e2af','#f38ba8'],
+  },
+  'catppuccin-latte': {
+    label: 'Catppuccin Latte', emoji: '\u2615', dark: false,
+    bg: { base:'#eff1f5', surface:'#e6e9ef', elevated:'#dce0e8', overlay:'#ccd0da',
+          hover:'rgba(76,79,105,0.05)', active:'rgba(76,79,105,0.09)',
+          raised:'#bcc0cc', statusbar:'#e6e9ef', tree:'#eff1f5' },
+    border: { subtle:'rgba(76,79,105,0.08)', default:'rgba(76,79,105,0.13)',
+              strong:'rgba(76,79,105,0.22)', mid:'rgba(76,79,105,0.10)' },
+    text: { primary:'#4c4f69', secondary:'#5c5f77', muted:'#8c8fa1', faint:'#acafbe' },
+    accent: '#8839ef',
+    green:'#40a02b', red:'#d20f39', blue:'#1e66f5', purple:'#8839ef', yellow:'#df8e1d',
+    depths: ['#8839ef','#1e66f5','#40a02b','#df8e1d','#d20f39'],
+  },
+  'cyberpunk': {
+    label: 'Cyberpunk', emoji: '\u{1F916}', dark: true,
+    bg: { base:'#0d0221', surface:'#12032e', elevated:'#1a0535', overlay:'#20063e',
+          hover:'rgba(255,0,255,0.04)', active:'rgba(255,0,255,0.07)',
+          raised:'#260843', statusbar:'#0d0221', tree:'#0f0228' },
+    border: { subtle:'rgba(255,0,255,0.10)', default:'rgba(255,0,255,0.18)',
+              strong:'rgba(255,0,255,0.32)', mid:'rgba(255,0,255,0.14)' },
+    text: { primary:'#f0e6ff', secondary:'#c9a9ff', muted:'#7a5a99', faint:'#4a2e66' },
+    accent: '#ff00ff',
+    green:'#00ff9f', red:'#ff2d6b', blue:'#00e5ff', purple:'#bf00ff', yellow:'#ffe600',
+    depths: ['#ff00ff','#00e5ff','#ffe600','#00ff9f','#ff2d6b'],
+  },
+  'tokyo-night': {
+    label: 'Tokyo Night', emoji: '\u{1F5FC}', dark: true,
+    bg: { base:'#1a1b26', surface:'#16161e', elevated:'#1f2335', overlay:'#24283b',
+          hover:'rgba(192,202,245,0.04)', active:'rgba(192,202,245,0.07)',
+          raised:'#2a2b3d', statusbar:'#16161e', tree:'#1a1b26' },
+    border: { subtle:'rgba(192,202,245,0.07)', default:'rgba(192,202,245,0.12)',
+              strong:'rgba(192,202,245,0.20)', mid:'rgba(192,202,245,0.09)' },
+    text: { primary:'#c0caf5', secondary:'#a9b1d6', muted:'#565f89', faint:'#3b4261' },
+    accent: '#7aa2f7',
+    green:'#9ece6a', red:'#f7768e', blue:'#7aa2f7', purple:'#bb9af7', yellow:'#e0af68',
+    depths: ['#7aa2f7','#bb9af7','#9ece6a','#e0af68','#f7768e'],
+  },
+  'tokyo-storm': {
+    label: 'Tokyo Storm', emoji: '\u26C8\uFE0F', dark: true,
+    bg: { base:'#24283b', surface:'#1f2335', elevated:'#292e42', overlay:'#2f3549',
+          hover:'rgba(192,202,245,0.05)', active:'rgba(192,202,245,0.08)',
+          raised:'#343b55', statusbar:'#1f2335', tree:'#24283b' },
+    border: { subtle:'rgba(192,202,245,0.07)', default:'rgba(192,202,245,0.12)',
+              strong:'rgba(192,202,245,0.20)', mid:'rgba(192,202,245,0.09)' },
+    text: { primary:'#c0caf5', secondary:'#a9b1d6', muted:'#565f89', faint:'#3b4261' },
+    accent: '#bb9af7',
+    green:'#9ece6a', red:'#f7768e', blue:'#7aa2f7', purple:'#bb9af7', yellow:'#e0af68',
+    depths: ['#bb9af7','#7aa2f7','#9ece6a','#e0af68','#f7768e'],
+  },
+  'dracula': {
+    label: 'Dracula', emoji: '\u{1F9DB}', dark: true,
+    bg: { base:'#282a36', surface:'#21222c', elevated:'#2d2f3f', overlay:'#343746',
+          hover:'rgba(248,248,242,0.04)', active:'rgba(248,248,242,0.07)',
+          raised:'#3a3c4e', statusbar:'#21222c', tree:'#282a36' },
+    border: { subtle:'rgba(98,114,164,0.18)', default:'rgba(98,114,164,0.28)',
+              strong:'rgba(98,114,164,0.45)', mid:'rgba(98,114,164,0.22)' },
+    text: { primary:'#f8f8f2', secondary:'#cfcfdf', muted:'#6272a4', faint:'#44475a' },
+    accent: '#bd93f9',
+    green:'#50fa7b', red:'#ff5555', blue:'#8be9fd', purple:'#bd93f9', yellow:'#f1fa8c',
+    depths: ['#bd93f9','#8be9fd','#50fa7b','#f1fa8c','#ff5555'],
+  },
+  'solarized-dark': {
+    label: 'Solarized Dark', emoji: '\u{1F305}', dark: true,
+    bg: { base:'#002b36', surface:'#073642', elevated:'#0d4050', overlay:'#08424f',
+          hover:'rgba(131,148,150,0.07)', active:'rgba(131,148,150,0.12)',
+          raised:'#13515f', statusbar:'#073642', tree:'#002b36' },
+    border: { subtle:'rgba(131,148,150,0.12)', default:'rgba(131,148,150,0.20)',
+              strong:'rgba(131,148,150,0.35)', mid:'rgba(131,148,150,0.16)' },
+    text: { primary:'#fdf6e3', secondary:'#eee8d5', muted:'#657b83', faint:'#586e75' },
+    accent: '#268bd2',
+    green:'#859900', red:'#dc322f', blue:'#268bd2', purple:'#6c71c4', yellow:'#b58900',
+    depths: ['#268bd2','#2aa198','#859900','#b58900','#dc322f'],
+  },
+  'solarized-light': {
+    label: 'Solarized Light', emoji: '\u2600\uFE0F', dark: false,
+    bg: { base:'#fdf6e3', surface:'#eee8d5', elevated:'#e8e2cf', overlay:'#ddd8c6',
+          hover:'rgba(101,123,131,0.06)', active:'rgba(101,123,131,0.10)',
+          raised:'#d5cfc0', statusbar:'#eee8d5', tree:'#fdf6e3' },
+    border: { subtle:'rgba(101,123,131,0.10)', default:'rgba(101,123,131,0.16)',
+              strong:'rgba(101,123,131,0.28)', mid:'rgba(101,123,131,0.13)' },
+    text: { primary:'#657b83', secondary:'#586e75', muted:'#839496', faint:'#93a1a1' },
+    accent: '#268bd2',
+    green:'#859900', red:'#dc322f', blue:'#268bd2', purple:'#6c71c4', yellow:'#b58900',
+    depths: ['#268bd2','#2aa198','#859900','#b58900','#dc322f'],
+  },
+  'nord': {
+    label: 'Nord', emoji: '\u2744\uFE0F', dark: true,
+    bg: { base:'#2e3440', surface:'#292e39', elevated:'#3b4252', overlay:'#424b5a',
+          hover:'rgba(216,222,233,0.05)', active:'rgba(216,222,233,0.08)',
+          raised:'#4c566a', statusbar:'#292e39', tree:'#2e3440' },
+    border: { subtle:'rgba(216,222,233,0.08)', default:'rgba(216,222,233,0.13)',
+              strong:'rgba(216,222,233,0.22)', mid:'rgba(216,222,233,0.10)' },
+    text: { primary:'#eceff4', secondary:'#d8dee9', muted:'#81909c', faint:'#4c566a' },
+    accent: '#88c0d0',
+    green:'#a3be8c', red:'#bf616a', blue:'#81a1c1', purple:'#b48ead', yellow:'#ebcb8b',
+    depths: ['#88c0d0','#81a1c1','#a3be8c','#ebcb8b','#bf616a'],
+  },
+  'gruvbox-dark': {
+    label: 'Gruvbox Dark', emoji: '\u{1F7E4}', dark: true,
+    bg: { base:'#1d2021', surface:'#282828', elevated:'#32302f', overlay:'#3c3836',
+          hover:'rgba(235,219,178,0.05)', active:'rgba(235,219,178,0.08)',
+          raised:'#504945', statusbar:'#282828', tree:'#1d2021' },
+    border: { subtle:'rgba(235,219,178,0.08)', default:'rgba(235,219,178,0.14)',
+              strong:'rgba(235,219,178,0.24)', mid:'rgba(235,219,178,0.11)' },
+    text: { primary:'#ebdbb2', secondary:'#d5c4a1', muted:'#928374', faint:'#665c54' },
+    accent: '#fabd2f',
+    green:'#b8bb26', red:'#fb4934', blue:'#83a598', purple:'#d3869b', yellow:'#fabd2f',
+    depths: ['#fabd2f','#83a598','#b8bb26','#d3869b','#fb4934'],
+  },
+  'one-dark': {
+    label: 'One Dark Pro', emoji: '\u26AB', dark: true,
+    bg: { base:'#21252b', surface:'#282c34', elevated:'#2c313a', overlay:'#323842',
+          hover:'rgba(171,178,191,0.05)', active:'rgba(171,178,191,0.08)',
+          raised:'#3e4451', statusbar:'#21252b', tree:'#21252b' },
+    border: { subtle:'rgba(171,178,191,0.08)', default:'rgba(171,178,191,0.14)',
+              strong:'rgba(171,178,191,0.24)', mid:'rgba(171,178,191,0.11)' },
+    text: { primary:'#abb2bf', secondary:'#9da5b4', muted:'#5c6370', faint:'#3e4451' },
+    accent: '#61afef',
+    green:'#98c379', red:'#e06c75', blue:'#61afef', purple:'#c678dd', yellow:'#e5c07b',
+    depths: ['#61afef','#c678dd','#98c379','#e5c07b','#e06c75'],
+  },
+  'github-dark': {
+    label: 'GitHub Dark', emoji: '\u{1F419}', dark: true,
+    bg: { base:'#0d1117', surface:'#161b22', elevated:'#1c2128', overlay:'#21262d',
+          hover:'rgba(240,246,252,0.05)', active:'rgba(240,246,252,0.08)',
+          raised:'#30363d', statusbar:'#161b22', tree:'#0d1117' },
+    border: { subtle:'rgba(240,246,252,0.07)', default:'rgba(240,246,252,0.12)',
+              strong:'rgba(240,246,252,0.20)', mid:'rgba(240,246,252,0.09)' },
+    text: { primary:'#f0f6fc', secondary:'#c9d1d9', muted:'#6e7681', faint:'#30363d' },
+    accent: '#58a6ff',
+    green:'#3fb950', red:'#f85149', blue:'#58a6ff', purple:'#d2a8ff', yellow:'#e3b341',
+    depths: ['#58a6ff','#d2a8ff','#3fb950','#e3b341','#f85149'],
+  },
+  'github-light': {
+    label: 'GitHub Light', emoji: '\u{1F419}', dark: false,
+    bg: { base:'#ffffff', surface:'#f6f8fa', elevated:'#ffffff', overlay:'#f0f2f4',
+          hover:'rgba(31,35,40,0.04)', active:'rgba(31,35,40,0.07)',
+          raised:'#eaeef2', statusbar:'#f6f8fa', tree:'#ffffff' },
+    border: { subtle:'rgba(31,35,40,0.08)', default:'rgba(31,35,40,0.12)',
+              strong:'rgba(31,35,40,0.20)', mid:'rgba(31,35,40,0.10)' },
+    text: { primary:'#1f2328', secondary:'#636c76', muted:'#9198a1', faint:'#adbac7' },
+    accent: '#0969da',
+    green:'#1a7f37', red:'#cf222e', blue:'#0969da', purple:'#8250df', yellow:'#9a6700',
+    depths: ['#0969da','#8250df','#1a7f37','#9a6700','#cf222e'],
+  },
+  'monokai': {
+    label: 'Monokai', emoji: '\u{1F3A8}', dark: true,
+    bg: { base:'#1e1f1c', surface:'#272822', elevated:'#2e2f2a', overlay:'#35362f',
+          hover:'rgba(248,248,242,0.04)', active:'rgba(248,248,242,0.07)',
+          raised:'#3e3d32', statusbar:'#1e1f1c', tree:'#1e1f1c' },
+    border: { subtle:'rgba(248,248,242,0.07)', default:'rgba(248,248,242,0.12)',
+              strong:'rgba(248,248,242,0.22)', mid:'rgba(248,248,242,0.09)' },
+    text: { primary:'#f8f8f2', secondary:'#cfcfc2', muted:'#75715e', faint:'#49483e' },
+    accent: '#a6e22e',
+    green:'#a6e22e', red:'#f92672', blue:'#66d9e8', purple:'#ae81ff', yellow:'#e6db74',
+    depths: ['#a6e22e','#66d9e8','#ae81ff','#e6db74','#f92672'],
+  },
+  'synthwave': {
+    label: "Synthwave '84", emoji: '\u{1F306}', dark: true,
+    bg: { base:'#1a1333', surface:'#221944', elevated:'#2a1f52', overlay:'#312660',
+          hover:'rgba(255,120,220,0.05)', active:'rgba(255,120,220,0.08)',
+          raised:'#3d2d72', statusbar:'#1a1333', tree:'#1a1333' },
+    border: { subtle:'rgba(255,120,220,0.10)', default:'rgba(255,120,220,0.18)',
+              strong:'rgba(255,120,220,0.30)', mid:'rgba(255,120,220,0.13)' },
+    text: { primary:'#ffffff', secondary:'#e2c9ff', muted:'#9d6eb5', faint:'#5d3e7a' },
+    accent: '#ff79c6',
+    green:'#72f1b8', red:'#fe4450', blue:'#36f9f6', purple:'#ff7edb', yellow:'#fede5d',
+    depths: ['#ff79c6','#36f9f6','#72f1b8','#fede5d','#fe4450'],
+  },
+  'material-ocean': {
+    label: 'Material Ocean', emoji: '\u{1F30A}', dark: true,
+    bg: { base:'#0f111a', surface:'#090b10', elevated:'#171c28', overlay:'#1b2030',
+          hover:'rgba(198,212,255,0.04)', active:'rgba(198,212,255,0.07)',
+          raised:'#222738', statusbar:'#090b10', tree:'#0f111a' },
+    border: { subtle:'rgba(198,212,255,0.07)', default:'rgba(198,212,255,0.12)',
+              strong:'rgba(198,212,255,0.20)', mid:'rgba(198,212,255,0.09)' },
+    text: { primary:'#8f93a2', secondary:'#717cb4', muted:'#464b5d', faint:'#2e3248' },
+    accent: '#82aaff',
+    green:'#c3e88d', red:'#f07178', blue:'#82aaff', purple:'#c792ea', yellow:'#ffcb6b',
+    depths: ['#82aaff','#c792ea','#c3e88d','#ffcb6b','#f07178'],
+  },
+};
+
+const ACCENT_SWATCHES = [
+  { id:'amber',  hex:'#f0b429', label:'Amber'  },
+  { id:'teal',   hex:'#2dd4bf', label:'Teal'   },
+  { id:'violet', hex:'#a78bfa', label:'Violet' },
+  { id:'rose',   hex:'#fb7185', label:'Rose'   },
+  { id:'sky',    hex:'#38bdf8', label:'Sky'    },
+  { id:'lime',   hex:'#a3e635', label:'Lime'   },
+  { id:'orange', hex:'#fb923c', label:'Orange' },
+  { id:'pink',   hex:'#f472b6', label:'Pink'   },
+];
+
+export { FULL_THEMES, ACCENT_SWATCHES };
